+++ src/sdk.ts
@@ -1,0 +1,6 @@
+/**
+ * Exports the API for the `@antv/mcp-server-chart` package.
+ * This includes chart generation capabilities and utility functions.
+ */
+export * from "./charts";
+export { callTool } from "./utils";
