--- src/utils/constants.ts
@@ -1,18 +1,0 @@
-// Chart types mapping
-export const ChartTypeMapping = {
-  generate_line_chart: "line",
-  generate_column_chart: "column",
-  generate_area_chart: "area",
-  generate_pie_chart: "pie",
-  generate_bar_chart: "bar",
-  generate_histogram_chart: "histogram",
-  generate_scatter_chart: "scatter",
-  generate_word_cloud_chart: "word-cloud",
-  generate_radar_chart: "radar",
-  generate_treemap_chart: "treemap",
-  generate_dual_axes_chart: "dual-axes",
-  generate_mind_map: "mind-map",
-  generate_network_graph: "network-graph",
-  generate_flow_diagram: "flow-diagram",
-  generate_fishbone_diagram: "fishbone-diagram",
-} as const;
